Start abstracting JavaScript-code for load datasets card
