+++ app/views/mpa/analyze.html.erb
@@ -226,7 +226,6 @@
   var selectedDatasets = <%= raw(@selected_datasets) %>;
 
   $(function () {
-      console.log(selectedDatasets);
       new MPA(selectedDatasets);
   });
 </script>
