+++ app/views/shared/_load_datasets_card.html.erb
@@ -1,3 +1,4 @@
+<%= javascript_pack_tag 'mpa' %>
 <div class='card card-nav datasets' style="height: 100%;">
   <div class='card-title card-title-colored'>
     <ul class="nav nav-tabs">
@@ -12,7 +13,7 @@
       <div class="tab-pane active" id="create">
         <%= render :partial => "shared/create_dataset_form", :locals => {:id_prefix => ""} %>
         <div class='search-buttons-centered'>
-          <%= button_tag(raw("<span class='glyphicon glyphicon-arrow-left grow'></span> Add to selected datasets"), { :name => 'add', :id => 'add_dataset_button', :class => "btn btn-default btn-animate input-item" }) %>
+          <%= button_tag(raw("<span class='glyphicon glyphicon-arrow-left grow'></span> Add to selected datasets"), { :name => 'add', :id => 'add_dataset_button', :class => "btn btn-default btn-animate load-datasets-input-item" }) %>
         </div>
       </div>
       <div class="tab-pane" id="database">
@@ -23,15 +24,15 @@
         <p>You can easily load data from the <a href="http://www.ebi.ac.uk/pride/" target="_blank">PRIDE</a> data repository. Simply enter an assay id (e.g. 8500) in the field below and click the 'Load PRIDE Dataset' button. The corresponding dataset will then be fetched using the PRIDE API and loaded into the search form on the left.</p>
         <div class="form-group">
           <label for="pride_exp_id">Assay id</label>
-          <%= text_field_tag('pride_exp_id', nil, :placeholder => 'e.g. 8500', :class => "form-control input-item")%>
+          <%= text_field_tag('pride_exp_id', nil, :placeholder => 'e.g. 8500', :class => "form-control load-datasets-input-item")%>
         </div>
         <%= render 'shared/pride_progress' %>
         <div class='search-buttons-centered'>
-          <%= button_tag(raw("Fetch PRIDE dataset"), { :type => "button", :class => "load-pride btn btn-default input-item" , :"data-loading-text" => "Loading..."}) %>
+          <%= button_tag(raw("Fetch PRIDE dataset"), { :type => "button", :class => "load-pride btn btn-default load-datasets-input-item" , :"data-loading-text" => "Loading..."}) %>
         </div>
         <%= render :partial => "shared/create_dataset_form", :locals => {:id_prefix => "pride-"} %>
         <div class="search-buttons-centered">
-          <%= button_tag(raw("<span class='glyphicon glyphicon-arrow-left grow'></span> Add to selected datasets"), { :name => 'add', :id => 'add_pride_dataset_button', :class => "btn btn-default btn-animate input-item" }) %>
+          <%= button_tag(raw("<span class='glyphicon glyphicon-arrow-left grow'></span> Add to selected datasets"), { :name => 'add', :id => 'add_pride_dataset_button', :class => "btn btn-default btn-animate load-datasets-input-item" }) %>
         </div>
       </div>
       <div class="tab-pane" id="local-storage">
